fix: fix absolute path to file

+++ src/utils/fs.js
@@ -1,6 +1,6 @@
 import fs from 'node:fs';
 import path from 'path';
-import { showCurrentDir } from './helpers.js';
+import { showCurrentDir, checkPath } from './helpers.js';
 
 export async function createFile(fileName) {
 
@@ -9,10 +9,16 @@ export async function createFile(fileName) {
     return
   }
 
-  const currentDir = process.cwd();
-  const PATH = path.join(currentDir, fileName);
+  const PATH = path.resolve(fileName);
 
   try {
+    const isExist = await checkPath(PATH);
+
+    if(isExist) {
+      console.error('The file with same name is already exists.');
+      return;
+    }
+
     await fs.promises.writeFile(PATH, '');
     console.log(`The file ${fileName} created!`);
   } catch  (err){
@@ -26,13 +32,25 @@ export async function renameFile(pathToFile, newName) {
 
   if(!pathToFile || !newName) {
     console.error('Operation failed: Please use the command format: rn path_to_file new_filename');
+    return;
   }
 
-  const currentDir = process.cwd();
-  const oldPath = path.join(currentDir, pathToFile);
-  const newPath = path.join(currentDir, newName);
+  const oldPath = path.resolve(pathToFile);
+  const newPath = path.resolve(newName);
 
   try {
+    const isExist = await checkPath(oldPath);
+    if(!isExist) {
+      console.error('There is no file with same name.');
+      return;
+    }
+  
+    const isSameExist = await checkPath(newPath);
+    if(isSameExist) {
+      console.error('The file with same name is already exists.');
+      return;
+    } 
+
     await fs.promises.rename(oldPath, newPath);
     console.log(`The file has been renamed to ${newName}.`)
   } catch(err) {
@@ -49,10 +67,15 @@ export async function deleteFile(fileName) {
     return
   }
 
-  const currentDir = process.cwd();
-  const PATH = path.join(currentDir, fileName);
+  const PATH = path.resolve(fileName);
 
   try {
+    const isExist = await checkPath(PATH);
+    if(!isExist) {
+      console.error('There is no file with this name.');
+      return;
+    }
+
     await fs.promises.unlink(PATH);
     console.log('File deleted.');
   } catch(err) {
